Rename static data file to be global
